omega_imagenes/requirements.txt
streamlit>=1.22
pillow
torch
torchvision
transformers
kornia
